requirements.txt
numpy==1.26.4
requests==2.32.3
watchdog==4.0.1
Pillow==10.4.0
rembg==2.0.59
onnxruntime==1.18.1